Remove console logs and add conditionals to fix consoleerrors
Add conditional logic so that event listeners are added to studentSelect
and subjectSelect only if these elements are present on the page. This solves
an error that appears in the console where an attempt is made to add an event
listener to a null object, being the case where the studentSelect and subjectSelect
elements don't exist on the page. These elements only appear in the user's onboarding
form which is shown if they don't have students or subjects associated with their profile.